+++ apps/search/src/pages/api/webhooks/saleor/product_variant_updated.ts
@@ -1,7 +1,12 @@
-import { SaleorAsyncWebhook } from "@saleor/app-sdk/handlers/next";
-import { ProductEditedDocument, ProductEditedSubscription } from "../../../../../generated/graphql";
+import { NextWebhookApiHandler, SaleorAsyncWebhook } from "@saleor/app-sdk/handlers/next";
+import {
+  ProductVariantUpdated,
+  ProductVariantUpdatedDocument,
+} from "../../../../../generated/graphql";
 import { saleorApp } from "../../../../../saleor-app";
-import { handler } from "./_index";
+import { AlgoliaSearchProvider } from "../../../../lib/algolia/algoliaSearchProvider";
+import { getAlgoliaConfiguration } from "../../../../lib/algolia/getAlgoliaConfiguration";
+import { createDebug } from "../../../../lib/debug";
 
 export const config = {
   api: {
@@ -9,13 +14,43 @@ export const config = {
   },
 };
 
-export const webhookProductVariantUpdated = new SaleorAsyncWebhook<
-  ProductEditedSubscription["event"]
->({
+export const webhookProductVariantUpdated = new SaleorAsyncWebhook<ProductVariantUpdated>({
   webhookPath: "api/webhooks/saleor/product_variant_updated",
   asyncEvent: "PRODUCT_VARIANT_UPDATED",
   apl: saleorApp.apl,
-  subscriptionQueryAst: ProductEditedDocument,
+  subscriptionQueryAst: ProductVariantUpdatedDocument,
 });
+
+export const handler: NextWebhookApiHandler<ProductVariantUpdated> = async (req, res, context) => {
+  const debug = createDebug(`Webhook handler - ${webhookProductVariantUpdated.asyncEvent}`);
+
+  const { event, authData } = context;
+  debug(
+    `New event ${event} (${context.payload?.__typename}) from the ${authData.domain} domain has been received!`
+  );
+
+  const { settings, errors } = await getAlgoliaConfiguration({ authData });
+
+  if (errors?.length || !settings) {
+    debug("Aborting due to lack of settings");
+    debug(errors);
+    return res.status(400).json({
+      message: errors[0].message,
+    });
+  }
+
+  const searchProvider = new AlgoliaSearchProvider({
+    appId: settings.appId,
+    apiKey: settings.secretKey,
+    indexNamePrefix: settings.indexNamePrefix,
+  });
+
+  const { productVariant } = context.payload;
+  if (productVariant) {
+    await searchProvider.updateProductVariant(productVariant);
+  }
+  res.status(200).end();
+  return;
+};
 
 export default webhookProductVariantUpdated.createHandler(handler);