+++ apps/data-importer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # saleor-app-data-importer
 
+## 1.1.3
+
+### Patch Changes
+
+- b874d10: Update @saleor/app-sdk to 0.29.0
+- Updated dependencies [648d99b]
+  - @saleor/apps-shared@1.1.1
+
+## 1.1.2
+
+### Patch Changes
+
+- 9f843b2: Update imports to @saleor/apps-shared
+- 9f843b2: Remove generated folders form git history
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+  - @saleor/apps-shared@1.1.0
+
 ## 1.1.1
 
 ### Patch Changes