+++ apps/invoices/CHANGELOG.md
@@ -1,5 +1,42 @@
 # saleor-app-invoices
 
+## 1.4.0
+
+### Minor Changes
+
+- b874d10: Add logging error in SaleorAsyncWebhook
+
+### Patch Changes
+
+- b874d10: Update @saleor/app-sdk to 0.29.0
+- Updated dependencies [648d99b]
+  - @saleor/apps-shared@1.1.1
+
+## 1.3.0
+
+### Minor Changes
+
+- c1dab0b: Unsticky app cards and change loader type
+
+## 1.2.0
+
+### Minor Changes
+
+- 0219561: Remove firstName and lastName fields in billing address of a company
+
+## 1.1.3
+
+### Patch Changes
+
+- 9f843b2: Update imports to @saleor/apps-shared
+- 9f843b2: Remove generated folders form git history
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+- Updated dependencies [9f843b2]
+  - @saleor/apps-shared@1.1.0
+
 ## 1.1.2
 
 ### Patch Changes