+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @saleor/apps-shared
 
+## 1.1.1
+
+### Patch Changes
+
+- 648d99b: Remove not needed console log
+
+## 1.1.0
+
+### Minor Changes
+
+- 9f843b2: Add main-bar and app-icon
+- 9f843b2: Make TitleBar sticky to the top
+- 9f843b2: Allow icon, text and theme to AppIcon
+- 9f843b2: Add root index file
+
+### Patch Changes
+
+- 9f843b2: Remove generated folders form git history
+
 ## 1.0.1
 
 ### Patch Changes